# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeFlow API
  version: "0.3"
  description: |
    Describe an app in plain English and get a working, running app back.

    **Auth.** `POST /api/v1/auth/signup` or `/login` returns a JWT. Send it as
    `Authorization: Bearer <token>` on everything else.

    **Errors** are RFC 9457 problem details (`application/problem+json`) with
    a `traceId` you can quote when reporting a problem.

    | Status | Means |
    |---|---|
    | 401 | not signed in |
    | 403 | you can see this project, but your role can't do that |
    | 404 | doesn't exist - **or** you have no access to it (indistinguishable on purpose) |
    | 402 | a plan limit (`quota`, `limit`, `used`, `plan` in the body) |
    | 409 | conflicts with current state (a reply already running, nothing to retry, ...) |
    | 429 | rate limited - see `Retry-After` |

    Every response carries `traceparent` and `X-Trace-Id` headers. Send a W3C
    `traceparent` and ForgeFlow continues your trace.

    This document is checked against the running application by
    `OpenApiContractTest`: an endpoint added without being documented here
    fails the build.
servers:
  - url: /
security:
  - bearer: []
...
paths:
...
  /api/v1/projects/{projectId}/chat/sessions:
    parameters: [{ $ref: "#/components/parameters/projectId" }]
    get:
      tags: [Chat]
      summary: List chat sessions (READ), most recent first
      responses:
        "200": { description: Sessions, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ChatSession" } } } } }
    post:
      tags: [Chat]
      summary: Create a chat session (WRITE)
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, properties: { title: { type: string, maxLength: 200, description: "Defaults to the first message" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/ChatSession" } } } }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}:
    parameters: [{ $ref: "#/components/parameters/projectId" }, { $ref: "#/components/parameters/sessionId" }]
    patch:
      tags: [Chat]
      summary: Rename a session (WRITE)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [title], properties: { title: { type: string, maxLength: 200 } } }
      responses:
        "200": { description: Renamed, content: { application/json: { schema: { $ref: "#/components/schemas/ChatSession" } } } }
    delete:
      tags: [Chat]
      summary: Delete a session (WRITE)
      responses:
        "204": { description: Deleted }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/messages:
    parameters: [{ $ref: "#/components/parameters/projectId" }, { $ref: "#/components/parameters/sessionId" }]
    get:
      tags: [Chat]
      summary: Load the full chat history (READ), oldest first
      responses:
        "200": { description: Messages, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ChatMessage" } } } } }
    post:
      tags: [Chat]
      summary: Send a message and wait for the reply (WRITE)
      requestBody: { $ref: "#/components/requestBodies/Message" }
      responses:
        "200": { description: The exchange, content: { application/json: { schema: { $ref: "#/components/schemas/ChatTurn" } } } }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "409": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/messages/stream:
    parameters: [{ $ref: "#/components/parameters/projectId" }, { $ref: "#/components/parameters/sessionId" }]
    post:
      tags: [Chat]
      summary: Chat stream - agent events, then a final `message` event (WRITE)
      requestBody: { $ref: "#/components/requestBodies/Message" }
      responses:
        "200": { $ref: "#/components/responses/AgentEvents" }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/messages/{messageId}/attachments/{attachmentId}:
    parameters:
      - { $ref: "#/components/parameters/projectId" }
      - { $ref: "#/components/parameters/sessionId" }
      - { name: messageId, in: path, required: true, schema: { type: integer, format: int64 } }
      - { name: attachmentId, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Chat]
      summary: An image sent with a message (READ)
      responses:
        "200": { description: The image, content: { image/png: { schema: { type: string, format: binary } }, image/jpeg: { schema: { type: string, format: binary } }, image/webp: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/messages/{messageId}/visual-review:
    parameters:
      - { $ref: "#/components/parameters/projectId" }
      - { $ref: "#/components/parameters/sessionId" }
      - { name: messageId, in: path, required: true, description: the assistant reply whose result is on screen, schema: { type: integer, format: int64 } }
    post:
      tags: [Chat]
      summary: AI checks its own app - a vision model reviews a screenshot of the preview (WRITE, metered)
      description: >
        The workbench screenshots the live preview in the browser right after a reply
        lands and sends it here. A vision model judges it against the conversation's
        requests. NEEDS_FIXES means at least one major issue; the workbench then sends
        one follow-up turn asking the agent to fix them. The latest review is returned
        on the reply in the message history.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [screenshot]
              properties:
                screenshot:
                  type: object
                  required: [mimeType, data]
                  properties:
                    mimeType: { type: string, enum: [image/png, image/jpeg, image/webp] }
                    data: { type: string, description: base64, no data-URL prefix; at most 4 MB decoded }
      responses:
        "200": { description: The verdict, content: { application/json: { schema: { $ref: "#/components/schemas/VisualReview" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/retry:
    parameters: [{ $ref: "#/components/parameters/projectId" }, { $ref: "#/components/parameters/sessionId" }]
    post:
      tags: [Chat]
      summary: Retry the last question - only if its reply failed (WRITE)
      responses:
        "200": { description: The retried exchange, content: { application/json: { schema: { $ref: "#/components/schemas/ChatTurn" } } } }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/projects/{projectId}/chat/sessions/{sessionId}/retry/stream:
    parameters: [{ $ref: "#/components/parameters/projectId" }, { $ref: "#/components/parameters/sessionId" }]
    post:
      tags: [Chat]
      summary: Retry, streamed (WRITE)
      responses:
        "200": { $ref: "#/components/responses/AgentEvents" }
        "409": { $ref: "#/components/responses/Problem" }

  # ----------------------------------------------------------------- files
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    projectId: { name: projectId, in: path, required: true, schema: { type: integer, format: int64 } }
    sessionId: { name: sessionId, in: path, required: true, schema: { type: integer, format: int64 } }
  requestBodies:
...
    Message:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [content]
            properties:
              content: { type: string, maxLength: 4000 }
              images:
                type: array
                maxItems: 3
                description: "Screenshot to app: up to 3 images the agent builds from (PNG/JPEG/WebP, 4 MB each, checked by magic bytes)"
                items:
                  type: object
                  required: [mimeType, data]
                  properties:
                    mimeType: { type: string, enum: [image/png, image/jpeg, image/webp] }
                    data: { type: string, description: "base64, no data: prefix" }
  responses:
    Problem:
      description: Problem details
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    QuotaExceeded:
      description: A plan limit was reached
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/QuotaProblem" } } }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until the next request will be accepted }
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    AgentEvents:
      description: |
        Server-Sent Events as the agent works: `status`, `thinking`, `tool`,
        `file`, `tool_failed`, `build`, `repair`, then `done` (the
        GenerateResponse) - or `error`.
        Chat streams end with a `message` event holding the saved ChatTurn.
      content: { text/event-stream: { schema: { type: string } } }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
    QuotaProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            quota: { type: string, enum: [PROJECTS, PREVIEWS, AI_TOKENS_PER_DAY] }
            limit: { type: integer, format: int64 }
            used: { type: integer, format: int64 }
            plan: { type: string }
            upgrade: { type: string }
...
    GenerateResponse:
      type: object
      properties:
        runId: { type: integer, format: int64 }
        status: { type: string, enum: [SUCCEEDED, FAILED] }
        stopReason: { type: string, enum: [FINISH_TOOL, NO_TOOL_CALL, MAX_REPAIRS, TIMEOUT, MAX_TOOL_CALLS, TOKEN_BUDGET, ERROR] }
        summary: { type: string }
        filesWritten: { type: array, items: { type: string } }
        toolCalls: { type: integer }
        repairRounds: { type: integer }
        buildPassed: { type: boolean }
        totalTokens: { type: integer }
        durationMs: { type: integer, format: int64 }
        errorMessage: { type: string }
        toolUsage:
          type: object
          description: Calls per tool name, e.g. {"edit_file":1,"finish":1,"read_file":1}
          additionalProperties: { type: integer }
...
    ChatSession:
      type: object
      properties:
        id: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        title: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChatMessage:
      type: object
      properties:
        id: { type: integer, format: int64 }
        sessionId: { type: integer, format: int64 }
        role: { type: string, enum: [user, assistant, system, tool] }
        content: { type: string }
        authorId: { type: integer, format: int64 }
        toolCalls:
          type: array
          items: { type: object, properties: { name: { type: string }, path: { type: string } } }
        tokensUsed: { type: integer }
        status: { type: string }
        runId: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        images:
          type: array
          description: images sent with a user message; bytes at .../messages/{id}/attachments/{imageId}
          items: { type: object, properties: { id: { type: integer, format: int64 }, mimeType: { type: string }, sizeBytes: { type: integer } } }
        review:
          nullable: true
          description: on assistant replies - the latest visual review of the result
          allOf: [{ $ref: "#/components/schemas/VisualReview" }]
    VisualReview:
      type: object
      properties:
        id: { type: integer, format: int64 }
        messageId: { type: integer, format: int64 }
        score: { type: integer, minimum: 1, maximum: 10 }
        verdict: { type: string, enum: [LOOKS_RIGHT, NEEDS_FIXES] }
        summary: { type: string }
        issues:
          type: array
          maxItems: 5
          items: { type: object, properties: { severity: { type: string, enum: [major, minor] }, text: { type: string } } }
        tokensUsed: { type: integer }
        createdAt: { type: string, format: date-time }
    ChatTurn:
      type: object
      properties:
        userMessage: { $ref: "#/components/schemas/ChatMessage" }
        assistantMessage: { $ref: "#/components/schemas/ChatMessage" }
        run: { $ref: "#/components/schemas/GenerateResponse" }
...